// src/types/schema.d.ts
export interface components {
    schemas: {
// ...
        TokenObtainPair: {
            email: string;
            password: string;
            readonly access: string;
            readonly refresh: string;
        };
        TokenRefresh: {
            readonly access: string;
            refresh: string;
        };
        UserProfile: {
            readonly id: number;
            /**
             * Email address
             * Format: email
             */
            readonly email: string;
            readonly role: components["schemas"]["RoleEnum"];
            phone_number?: string | null;
            readonly is_verified: boolean;
            readonly doctor_profile: components["schemas"]["AccountDoctorProfile"];
            readonly patient_profile: components["schemas"]["PatientProfile"];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}